Image_models/Open_world_models/OVID/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
torchaudio>=0.8.0
pytorch-lightning>=1.6.0
opencv-python>=4.5.0
Pillow>=8.0.0
imgaug>=0.4.0
numpy>=1.19.0
scipy>=1.6.0
einops>=0.3.0
ftfy>=6.0.0
regex>=2021.4.4
git+https://github.com/openai/CLIP.git
matplotlib>=3.3.0
tqdm>=4.62.0
pickle5>=0.0.11; python_version < "3.8"